requirements.txt
flask
flask-cors
requests
pymysql
pyautogui
keyboard
mouse
gunicorn
cryptography
pytz
flask-login
flask-login
bcrypt==4.3.0
matplotlib
mysql-connector-python
openai
reportlab
pandas
openpyxl
SQLAlchemy>=2.0
numpy